requirements.txt
aiohttp==3.7.3
alpha-vantage==2.2.0
astroid==2.4.2
async-timeout==3.0.1
attrs==20.3.0
certifi==2020.12.5
chardet==3.0.4
idna==2.10
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
multidict==5.1.0
pylint==2.6.0
PyYAML==5.3.1
requests==2.25.1
six==1.15.0
slacker==0.14.0
toml==0.10.2
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.26.2
wrapt==1.12.1
yarl==1.6.3
